Solves importance sampling bug
ParSamples is a Numpy array, not a GetDist object, so ParSample.H0 does not exist. I modified MCEvidence.py to extract the parameter names from the .paramnames file, and passed this to self.names. This allows the correct column of ParSamples (corresponding to H0) to be extracted and used in the importance sampling function (h0_gauss_lnp) in planck_mcevidence.py).